Refine keysizes empty-key handling, helper signature, and XADD test
- Update the kvsUpdateHistogram() comments that claimed only strings can
  be empty: streams can also be persistently empty (XGROUP CREATE ...
  MKSTREAM, or deleting all entries from a stream), so bin 0 for a stream
  is intentional, not accidental.

- Add a test covering XADD's append-then-trim-in-one-command path
  (MAXLEN 0 leaves a new stream at length 0 -> bin 0; later MAXLEN 1 keeps
  it at bin 1), locking down the single post-command histogram update.

- Make keysizesHistIdx() take uint32_t to match the object type field
  (object.h: type:4) and drop the (int) cast at its kvsUpdateHistogram
  call site. Out-of-range/untracked types still fall through to -1.

+++ src/db.c
@@ -90,7 +90,7 @@ void updateLRM(robj *o) {
  *               [0,1)->0
  */
 void kvsUpdateHistogram(keysizesHist kvstoreHist, uint32_t type, int64_t oldLen, int64_t newLen) {
-    int idx = keysizesHistIdx((int) type);
+    int idx = keysizesHistIdx(type);
     if (unlikely(idx < 0))
         return;
 
@@ -102,8 +102,8 @@ void kvsUpdateHistogram(keysizesHist kvstoreHist, uint32_t type, int64_t oldLen,
     } else {
         /* here, oldLen can be either 0 or -1 */
         if (oldLen == 0) {
-            /* Only strings can be empty. Yet, a command flow might temporarily
-             * dbAdd() empty collection, and only after add elements. */
+            /* Only strings and streams can be empty. Yet, a command flow might
+             * temporarily dbAdd() empty collection, and only after add elements. */
             kvstoreHist[idx][0]--;
             debugServerAssert(kvstoreHist[idx][0] >= 0);
         }
@@ -116,8 +116,8 @@ void kvsUpdateHistogram(keysizesHist kvstoreHist, uint32_t type, int64_t oldLen,
     } else {
         /* here, newLen can be either 0 or -1 */
         if (newLen == 0) {
-            /* Only strings can be empty. Yet, a command flow might temporarily
-             * dbAdd() empty collection, and only after add elements. */
+            /* Only strings and streams can be empty. Yet, a command flow might
+             * temporarily dbAdd() empty collection, and only after add elements. */
             kvstoreHist[idx][0]++;
         }
     }
